chore: add indexer service
