+++ apps/riven/lib/graphql/resolvers/vfs/vfs.resolver.ts
@@ -9,16 +9,16 @@ import { VfsEntryStat } from "./types/vfs-entry-stat.type.ts";
 export class VfsResolver {
   @Query(() => VfsEntryStat)
   vfsEntryStat(@Arg("path") path: string): Promise<VfsEntryStat> {
-    return database.vfs.getEntryStat(path);
+    return database.vfsService.getEntryStat(path);
   }
 
   @Query(() => MediaEntry, { nullable: true })
   vfsEntry(@Arg("path") path: string): Promise<MediaEntry | null> {
-    return database.vfs.getEntry(path);
+    return database.vfsService.getEntry(path);
   }
 
   @Query(() => [String])
   vfsDirectoryEntryPaths(@Arg("path") path: string): Promise<string[]> {
-    return database.vfs.getDirectoryEntryPaths(path);
+    return database.vfsService.getDirectoryEntryPaths(path);
   }
 }
