+++ apps/riven/lib/message-queue/flows/index-item/index-item.processor.ts
@@ -3,9 +3,8 @@ import { MediaItemIndexErrorIncorrectState } from "@repo/util-plugin-sdk/schemas
 
 import { UnrecoverableError } from "bullmq";
 
+import { database } from "../../../database/database.ts";
 import { requestIndexDataProcessorSchema } from "./index-item.schema.ts";
-import { persistMovieIndexerData } from "./utilities/persist-movie-indexer-data.ts";
-import { persistShowIndexerData } from "./utilities/persist-show-indexer-data.ts";
 
 import type { MediaItemIndexRequestedResponse } from "@repo/util-plugin-sdk/schemas/events/media-item.index.requested.event";
 
@@ -34,8 +33,8 @@ export const indexItemProcessor =
     try {
       const updatedItem =
         item.type === "movie"
-          ? await persistMovieIndexerData({ item })
-          : await persistShowIndexerData({ item });
+          ? await database.indexerService.indexMovie({ item })
+          : await database.indexerService.indexShow({ item });
 
       sendEvent({
         type: "riven.media-item.index.success",
