+++ apps/riven/lib/services/indexer/utilities/persist-movie-indexer-data.spec.ts
@@ -4,10 +4,10 @@ import { MediaItemIndexErrorIncorrectState } from "@repo/util-plugin-sdk/schemas
 import { DateTime } from "luxon";
 import { expect } from "vitest";
 
-import { it } from "../../../../__tests__/test-context.ts";
-import { persistMovieIndexerData } from "./persist-movie-indexer-data.ts";
+import { it } from "../../../__tests__/test-context.ts";
 
 it("returns the media item if processed successfully", async ({
+  indexerService,
   factories: { movieItemRequestFactory },
 }) => {
   const requestedId = "tt1234567";
@@ -17,17 +17,15 @@ it("returns the media item if processed successfully", async ({
     state: "requested",
   });
 
-  const result = await persistMovieIndexerData({
-    item: {
-      id: itemRequest.id,
-      title: "Test Movie",
-      imdbId: requestedId,
-      contentRating: "g",
-      genres: [],
-      type: "movie",
-      runtime: 40,
-      releaseDate: DateTime.now().toISO(),
-    },
+  const result = await indexerService.indexItem({
+    id: itemRequest.id,
+    title: "Test Movie",
+    imdbId: requestedId,
+    contentRating: "g",
+    genres: [],
+    type: "movie",
+    runtime: 40,
+    releaseDate: DateTime.now().toISO(),
   });
 
   expect(result).instanceOf(Movie);
@@ -41,6 +39,7 @@ it("returns the media item if processed successfully", async ({
 });
 
 it("throws a MediaItemIndexErrorIncorrectState error if the item request is in an incorrect state", async ({
+  indexerService,
   factories: { movieItemRequestFactory },
 }) => {
   const requestedId = "1234";
@@ -51,17 +50,15 @@ it("throws a MediaItemIndexErrorIncorrectState error if the item request is in a
   });
 
   await expect(
-    persistMovieIndexerData({
-      item: {
-        id: itemRequest.id,
-        title: "Test Movie",
-        imdbId: requestedId,
-        contentRating: "g",
-        genres: [],
-        type: "movie",
-        runtime: 40,
-        releaseDate: DateTime.now().toISO(),
-      },
+    indexerService.indexItem({
+      id: itemRequest.id,
+      title: "Test Movie",
+      imdbId: requestedId,
+      contentRating: "g",
+      genres: [],
+      type: "movie",
+      runtime: 40,
+      releaseDate: DateTime.now().toISO(),
     }),
   ).rejects.toThrow(MediaItemIndexErrorIncorrectState);
 });
